requirements.txt
openai>=1.40.0
PyYAML>=6.0
langchain-text-splitters>=0.3.0
